i did second and i hope i do everything right AAAAA

+++ org/nehyshta/solutions/first.java
@@ -8,9 +8,9 @@ public class first {
     public static void main(String[] args) {
         Scanner s = new Scanner(System.in);
 
-        float a = s.nextInt();
-        float b = s.nextInt();
-        float c = s.nextInt();
+        float a = s.nextFloat();
+        float b = s.nextFloat();
+        float c = s.nextFloat();
 
         float z = ((a-3)*b/2)+c;
         System.out.printf("z is %.0f", z);
